+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 ### Added
 
 - a lot more coercion cases allowed for node puts (non-source)
-  - `pattern`, `arg`, `alias`, `withitem`, `TypeVar` and `TypeVarTuple` to `expr`
+  - `pattern`, `arguments`, `arg`, `alias`, `withitem`, `TypeVar` and `TypeVarTuple` to `expr`
   - custom slice types `_Assign_targets`, `_decorator_list`, `_arglikes`, `_comprehension_ifs`, `_aliases`, `_withitems` and `_type_params` to `Tuple`
   - all sequence types, custom and standard, to custom slice types